Move PageTitle.java to /utils
