+++ app/src/main/java/fr/free/nrw/commons/LicenseList.java
@@ -13,6 +13,8 @@ import java.util.HashMap;
 import java.util.Locale;
 import java.util.Map;
 
+import fr.free.nrw.commons.utils.PageTitle;
+
 /**
  * Represents a list of Licenses
  */
