+++ app/src/main/java/fr/free/nrw/commons/auth/LoginActivity.java
@@ -38,7 +38,7 @@ import butterknife.BindView;
 import butterknife.ButterKnife;
 import butterknife.OnClick;
 import fr.free.nrw.commons.BuildConfig;
-import fr.free.nrw.commons.PageTitle;
+import fr.free.nrw.commons.utils.PageTitle;
 import fr.free.nrw.commons.R;
 import fr.free.nrw.commons.utils.Utils;
 import fr.free.nrw.commons.WelcomeActivity;
